pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai4a11y"
version = "0.1.0"
description = "AI agents that adapt the web to each user's abilities."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "AI for Accessibility Collective"},
]
keywords = ["accessibility", "agents", "wcag", "a11y", "assistive-technology"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "playwright>=1.40",
    "google-genai>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
anthropic = [
    "anthropic>=0.40",
]
openai = [
    "openai>=1.0",
]
dev = [
    "pytest>=8.0",
]

[project.scripts]
a11y = "ai4a11y.cli:main"

[tool.setuptools.packages.find]
include = ["ai4a11y*"]

[project.urls]
Homepage = "https://github.com/chuanenlin/AI-for-Accessibility-Toolkit-Draft"
Documentation = "https://github.com/chuanenlin/AI-for-Accessibility-Toolkit-Draft/blob/main/docs/architecture.md"
Repository = "https://github.com/chuanenlin/AI-for-Accessibility-Toolkit-Draft"
Issues = "https://github.com/chuanenlin/AI-for-Accessibility-Toolkit-Draft/issues"
